fix warnings (#2360)
* fix warnings

* fix warnings - phase 2

* remove exclude node_modules dari swiftlint.yml

* fix possible bug where the parameter passed is not boolean

* remove unneeded code, add try, remove optionals, remove force unwraps

* pod files

* remove commented codes, add discardableResult, remove force cast

* fix chain optional

* remove unnecessary parentheses and void casting

* fix build error

+++ ios/Tokopedia/Features/UIRefactor/Alert/AlertDatePickerView.m
@@ -132,10 +132,10 @@
                 NSDateFormatter *dateFormat = [[NSDateFormatter alloc] init];
                 [dateFormat setDateFormat:@"YYYY-MM-dd"];
 
-                NSString *maximum = [NSString stringWithFormat:@"%ld-12-31", maximumYear];
+                NSString *maximum = [NSString stringWithFormat:@"%ld-12-31", (long)maximumYear];
                 NSDate *maximumDate = [dateFormat dateFromString:maximum];
 
-                NSString *minimum = [NSString stringWithFormat:@"%ld-01-01", minimumYear];
+                NSString *minimum = [NSString stringWithFormat:@"%ld-01-01", (long)minimumYear];
                 NSDate *minimumDate = [dateFormat dateFromString:minimum];
 
                 [_datepicker setMaximumDate:maximumDate];
